Fix problem where node eventloop stops after move.
When node moves it will send no more timer events to itself if there are no more messages in queue. This fixes that. Also adds a printout when the unsetPeer() is called.

Furthermore there seems to be a complete clusterfuck of spaghetti in everything related to BindUpdates so added some TODOs to remember to talk to Viking about this.